% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grainsize_transform.R
\name{grainsize_transform}
\alias{grainsize_transform}
\title{grainsize_transform}
\usage{
grainsize_transform(x)
}
\arguments{
\item{x}{.CSV file using the template included with this package}
}
\value{
returns a list with df_head and df_data for the header and data
information.
}
\description{
grainsize_transform
}
\examples{
grainsize_transform(file)
}
