\name{grainsizeProc}
\alias{grain size processing}
\title{Sediment grainsize processing script}
\usage{
grainsize_transform(file)
}
\description{
Converts the standard grainsize template from Dr. Brent Law's sediment grain size (SGS) laboratory. There is an example dataset included in the 'data' folder included with this repository.
}
\examples{
grainsize_transform('data/example.csv')
}
